requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dotenv==0.9.9
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.19.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dotenv==1.1.0
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
setuptools==78.1.1
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
wheel==0.45.1
yt-dlp==2025.5.22
